+++ api/src/main/java/org/allaymc/api/entity/component/player/EntityPlayerDebugShapeViewerComponent.java
@@ -1,0 +1,10 @@
+package org.allaymc.api.entity.component.player;
+
+import org.allaymc.api.debugshape.DebugShapeViewer;
+import org.allaymc.api.entity.component.EntityComponent;
+
+/**
+ * @author daoge_cmd
+ */
+public interface EntityPlayerDebugShapeViewerComponent extends EntityComponent, DebugShapeViewer {
+}