+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - (API) Introduced field `LiquidHardenEvent.hardenedBlockPosition` which is the position of the hardened block.
 - (API) Introduced `api_version` for plugin descriptor, which can set the api version requirement of a plugin.
 - (API) Introduced component `EntityPlayerScoreboardViewerComponent` for player.
+- (API) Introduced component `EntityPlayerDebugShapeViewerComponent` for player.
 - Introduced dirty flag for block layers in chunk section. Now blocks will only be rewritten to the database if they are changed. This
   would speed up the time used during server shutdown significantly if there are many only loaded chunks.
 - Added support for basic multi-version. The server now support 1.21.80 - 1.21.100 client to join.